\name{readRWS}
\alias{readRWS}
\alias{RWSstations}

\title{
  Reads data files from the Dutch governmental organisation RWS (rijkswaterstaat).
  Only suitable for biogeochemical variables, and water height and temperature.
}

\description{
  \code{readRWS} reads text files with water data generated by the Dutch governmental organisation RWS. 
  The dutch is translated into english, and the units are converted to more standard units.

  \code{RWSstations} contains the positions and names of the RWS stations that are present in the Waddensea and that contain biogeochemical data.
}

\usage{
readRWS (file, dir="", attr=NULL, format="wide", ...)

data(RWSstations)
}

\arguments{
  \item{file }{The name of the file(s) to read. The name(s) will be kept as an attribute of the output. 
  }
  \item{dir }{The name of the directory holding the file(s). 
  }
  \item{attr }{A list with information to be added to the attributes of the output.
  }
  \item{format }{If not equal to "wide", will return it in long format. See value.
  }
  \item{... }{Any other arguments passed to the reading function , i.e. dec, quote, fill, etc.
  }
}
\note{
Several small example files are in the package's "extdata" directory: 

\itemize{
 \item for RWS data: "RWSdatasmall.csv" and "RWSdatasmall2.csv"
}

The package directory can be grasped with function \code{system.file}. See examples.

}
\note{
Information on the RWS stations that are located in the Waddensea can be found 
in dataset \code{RWSstations}.
}

\value{
  Function \code{readRWS} returns a data.frame that will be different depending on the settings of \code{format}
  
  \itemize{

  \item If the argument \code{format} is equal to \code{"wide"} then the data are in wide format, and the \code{data.frame} has the following columns:

 \itemize{
  \item station: the identifier of the station (acronym)
  \item longitude, latitude: the coordinates of the station
  \item datetime: the date and time of the sampling
  \item ... : the measured variables. 
 }
 Units of the measured variables can be found with \code{attributes(object)$variables}. Typical units are mmol/m3, m, mg/L.  
  
  \item  If the argument \code{format} is not equal to \code{"wide"} then the data are in long format, and 
the data.frame has the following columns:

 \itemize{
  \item station: the identifier of the station (acronym)
  \item longitude, latitude: the coordinates of the station
  \item datetime: the date and time of the sampling
  \item variable: the name of what has been measured 
  \item code: the short name of what has been measured 
  \item value: the measured value
  \item unit: the unit of the measurement
  \item vnr: referring to the same value in the \code{variables} attributes 
 }

}
  
The attributes of the object will contain: 
\itemize{
 \item source: the name of the originating organization, 
 \item file: the name of the file from which the data were read,
 \item processing: text keeping track of processing of the data,
 \item variables: code, name, property, method, unit, sensor used, time from and to, and the number referring to column vnr in the data, 
 \item stations: the long names, acronyms, and position of the stations, 
 \item EPSG: the coordinate system (EPSG nr) of the positions,
 \item processing: when the dataset was generated, and other operations performed on the data.
}

The attributes of the object also contain all information passed with argument "attr".
}

\examples{

# path of the external data of the package
fpath <- system.file("extdata", package="dtLife")

dir(fpath)  # shows all files 

## =============================================================
## Read the RWS file
## =============================================================

RWS <- readRWS(file = "RWSdatasmall.csv", 
               dir  = fpath)

head(RWS, n=2)

# metadata
AF1 <- meta(RWS)

AF1$variables
AF1$stations
AF1$processing

## =============================================================
## Read the files, add extra information to table and attributes
## =============================================================

RWS2 <- readRWS(file = "RWSdatasmall2.csv", 
                dir  = fpath,
                attr = list(author = "karline soetaert", project = "LTER-life"))

AF2 <- meta(RWS2)
AF2$author

## =============================================================
## Read and combine the two files
## =============================================================

RWS3L <- readRWS(file = c("RWSdatasmall.csv", "RWSdatasmall2.csv"), 
                 dir  = fpath,
                 format = "long"
                )
head(RWS3L)

RWS3 <- readRWS(file = c("RWSdatasmall.csv", "RWSdatasmall2.csv"), 
                dir  = fpath,
                format = "wide"
                )
head(RWS3)

## =============================================================
## Show the positions of the RWS stations
## =============================================================

with(RWSstations, 
   plot(x   = longitude, 
        y   = latitude, 
        pch = 18, cex = 4, las = 1))
   
with(RWSstations, 
   text(x      = longitude+0.01, 
        y      = latitude+0.01, 
        labels = station, adj = 0))

plot(Shape, add = TRUE)   

}
\references{
https://waterinfo.rws.nl/
}
\seealso{
\link{readKNMI}, \link{readBathymetry} for other reading functions.
}

\author{
  Karline Soetaert <karline.soetaert@nioz.nl>
}
\details{
The relevant columns of the dutch files mean (taken from the first part of the text files):
\itemize{
 \item "MEETPUNT_IDENTIFICATIE" ,  "LOCATIE_CODE": name of the station and code           
 \item "PARAMETER_OMSCHRIJVING", "PARAMETER_.CODE": name of the variable and code                    
 \item "GROOTHEID_OMSCHRIJVING", "GROOTHEID_.CODE": description of the unit of the variable, and its code
 \item "EENHEID_CODE" : unit of the variable
 \item "COMPARTIMENT_OMSCHRIJVING",  "COMPARTIMENT_CODE": where the sample was taken                                                
 \item "WAARDEBEPALINGSMETHODE_OMSCHRIJVING", "WAARDEBEPALINGSMETHODE_CODE": method of measurement
 \item "BEMONSTERINGSSOORT_OMSCHRIJVING", "BEMONSTERINGSSOORT_CODE" : how the sample was taken            
 \item "WAARNEMINGDATUM",  "WAARNEMINGTIJD..MET.CET." : date and time of sampling
 \item  "NUMERIEKEWAARDE", "ALFANUMERIEKEWAARDE"  : value                                      
 \item "KWALITEITSOORDEEL_CODE", "STATUSWAARDE" : quality valuation              
 \item "OPDRACHTGEVENDE_INSTANTIE" : sponsor                     
 \item "BEMONSTERINGSAPPARAAT_OMSCHRIJVING",  "BEMONSTERINGSAPPARAAT_CODE" : device with which the sample was taken       
 \item  "BEMONSTERINGSHOOGTE",  "REFERENTIEVLAK" : height of sampling and reference level
 \item "X", "Y", "EPSG"    : x-, y- coordinates and coordinate system
 }         
}
\keyword{utilities}
